requirements.txt
pycryptodome
secretsharing
tkinter  
